Move CompiledContract and CompiledContractCache to near-vm-errors (#9176)
Their actual location should be near-vm-runtime. However, near-vm-runtime currently depends on near-primitives, and near-primitives needs to have access to CompiledContract and CompiledContractCache. So it would not be easy to move them to near-vm-runtime right now.

However, the end-goal of this refactoring is actually to be able to merge near-vm-errors and near-vm-runtime. So even if we move them to the "wrong" place right now, it will become the "right" place as soon as the other refactorings necessary for that merge to happen will be complete.
