+++ core/primitives/src/runtime/apply_state.rs
@@ -2,9 +2,10 @@ use crate::runtime::migration_data::{MigrationData, MigrationFlags};
 use crate::{
     hash::CryptoHash,
     runtime::config::RuntimeConfig,
-    types::{Balance, BlockHeight, CompiledContractCache, EpochHeight, EpochId, Gas},
+    types::{Balance, BlockHeight, EpochHeight, EpochId, Gas},
     version::ProtocolVersion,
 };
+use near_vm_errors::CompiledContractCache;
 use std::sync::Arc;
 
 #[derive(Debug)]
