+++ core/primitives/src/types.rs
@@ -936,21 +936,6 @@ pub enum TransactionOrReceiptId {
     Receipt { receipt_id: CryptoHash, receiver_id: AccountId },
 }
 
-#[derive(Debug, Clone, PartialEq, BorshDeserialize, BorshSerialize)]
-pub enum CompiledContract {
-    CompileModuleError(near_vm_errors::CompilationError),
-    Code(Vec<u8>),
-}
-
-/// Cache for compiled modules
-pub trait CompiledContractCache: Send + Sync {
-    fn put(&self, key: &CryptoHash, value: CompiledContract) -> std::io::Result<()>;
-    fn get(&self, key: &CryptoHash) -> std::io::Result<Option<CompiledContract>>;
-    fn has(&self, key: &CryptoHash) -> std::io::Result<bool> {
-        self.get(key).map(|entry| entry.is_some())
-    }
-}
-
 /// Provides information about current epoch validators.
 /// Used to break dependency between epoch manager and runtime.
 pub trait EpochInfoProvider {
