+++ core/primitives/src/utils.rs
@@ -11,7 +11,7 @@ use serde;
 use crate::hash::{hash, CryptoHash};
 use crate::receipt::Receipt;
 use crate::transaction::SignedTransaction;
-use crate::types::{CompiledContractCache, NumSeats, NumShards, ShardId};
+use crate::types::{NumSeats, NumShards, ShardId};
 use crate::version::{
     ProtocolVersion, CORRECT_RANDOM_VALUE_PROTOCOL_VERSION, CREATE_HASH_PROTOCOL_VERSION,
     CREATE_RECEIPT_ID_SWITCH_TO_CURRENT_BLOCK_VERSION,
@@ -449,12 +449,6 @@ where
 {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "{:?}", serde_json::to_string(&self.0).unwrap())
-    }
-}
-
-impl fmt::Debug for dyn CompiledContractCache {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "Compiled contracts cache")
     }
 }
