+++ integration-tests/src/tests/client/process_blocks.rs
@@ -3584,9 +3584,9 @@ mod contract_precompilation_tests {
     use super::*;
     use near_primitives::contract::ContractCode;
     use near_primitives::test_utils::MockEpochInfoProvider;
-    use near_primitives::types::CompiledContractCache;
     use near_primitives::views::ViewApplyState;
     use near_store::{Store, StoreCompiledContractCache, TrieUpdate};
+    use near_vm_errors::CompiledContractCache;
     use near_vm_runner::get_contract_cache_key;
     use near_vm_runner::internal::VMKind;
     use node_runtime::state_viewer::TrieViewer;
