+++ nearcore/Cargo.toml
@@ -60,6 +60,7 @@ near-rosetta-rpc = { workspace = true, optional = true }
 near-store.workspace = true
 near-telemetry.workspace = true
 near-vm-runner.workspace = true
+near-vm-errors.workspace = true
 node-runtime.workspace = true
 near-config-utils.workspace = true
