+++ runtime/near-vm-runner/src/cache.rs
@@ -3,8 +3,7 @@ use crate::vm_kind::VMKind;
 use borsh::BorshSerialize;
 use near_primitives::contract::ContractCode;
 use near_primitives::hash::CryptoHash;
-use near_primitives::types::{CompiledContract, CompiledContractCache};
-use near_vm_errors::{CacheError, CompilationError};
+use near_vm_errors::{CacheError, CompilationError, CompiledContract, CompiledContractCache};
 use near_vm_logic::{ProtocolVersion, VMConfig};
 use std::collections::HashMap;
 use std::fmt;
