+++ runtime/near-vm-runner/src/runner.rs
@@ -3,9 +3,8 @@ use crate::vm_kind::VMKind;
 use near_primitives::config::VMConfig;
 use near_primitives::contract::ContractCode;
 use near_primitives::runtime::fees::RuntimeFeesConfig;
-use near_primitives::types::CompiledContractCache;
 use near_primitives::version::ProtocolVersion;
-use near_vm_errors::{CacheError, CompilationError, VMRunnerError};
+use near_vm_errors::{CacheError, CompilationError, CompiledContractCache, VMRunnerError};
 use near_vm_logic::types::PromiseResult;
 use near_vm_logic::{External, VMContext, VMOutcome};
