+++ runtime/near-vm-runner/src/tests/cache.rs
@@ -10,9 +10,8 @@ use assert_matches::assert_matches;
 use near_primitives::contract::ContractCode;
 use near_primitives::hash::CryptoHash;
 use near_primitives::runtime::fees::RuntimeFeesConfig;
-use near_primitives::types::{CompiledContract, CompiledContractCache};
 use near_stable_hasher::StableHasher;
-use near_vm_errors::VMRunnerError;
+use near_vm_errors::{CompiledContract, CompiledContractCache, VMRunnerError};
 use near_vm_logic::mocks::mock_external::MockedExternal;
 use near_vm_logic::VMConfig;
 use std::hash::{Hash, Hasher};
