+++ runtime/runtime-params-estimator/Cargo.toml
@@ -44,6 +44,7 @@ near-o11y.workspace = true
 near-primitives.workspace = true
 near-store.workspace = true
 near-test-contracts.workspace = true
+near-vm-errors.workspace =  true
 near-vm-logic.workspace = true
 near-vm-runner = { workspace = true, features = ["clap"] }
 nearcore.workspace = true
