+++ runtime/runtime-params-estimator/src/function_call.rs
@@ -3,8 +3,9 @@ use crate::gas_cost::{GasCost, LeastSquaresTolerance};
 use crate::vm_estimator::create_context;
 use near_primitives::contract::ContractCode;
 use near_primitives::runtime::config_store::RuntimeConfigStore;
-use near_primitives::types::{CompiledContractCache, ProtocolVersion};
+use near_primitives::types::ProtocolVersion;
 use near_store::StoreCompiledContractCache;
+use near_vm_errors::CompiledContractCache;
 use near_vm_logic::mocks::mock_external::MockedExternal;
 use near_vm_runner::internal::VMKind;
 use std::fmt::Write;
