+++ runtime/runtime-params-estimator/src/vm_estimator.rs
@@ -4,9 +4,9 @@ use crate::{utils::read_resource, REAL_CONTRACTS_SAMPLE};
 use near_primitives::contract::ContractCode;
 use near_primitives::hash::CryptoHash;
 use near_primitives::runtime::config_store::RuntimeConfigStore;
-use near_primitives::types::{CompiledContract, CompiledContractCache};
 use near_primitives::version::PROTOCOL_VERSION;
 use near_store::StoreCompiledContractCache;
+use near_vm_errors::{CompiledContract, CompiledContractCache};
 use near_vm_logic::VMContext;
 use near_vm_runner::internal::VMKind;
